Reworked how categories/super restricted updates work
Renamed postSuperRestrictedWord to postUpdateWord as categories can now be updated.
Add function, haveCategoriesChanged, to RestrictedWordController which is used in postUpdateWord.
Reworked postUpdateWord when determining what has been updated as this will determine the call to the api.
Add new enum UpdateFields which is used in postUpdateWord.
Update word.html to use html rather than nunjucks govukTable. This in prep for using the tags from the filter work on the main page.

PAS-437